+++ examples/js.eslintrc.js
@@ -1,6 +1,6 @@
 // https://github.com/SrBrahma/eslint-config-gev
 // This is a workaround for https://github.com/eslint/eslint/issues/3458
-require('@rushstack/eslint-patch/modern-module-resolution');
+require("@rushstack/eslint-patch/modern-module-resolution")
 
 module.exports = {
   root: true,
@@ -8,7 +8,7 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ['eslint-config-gev/js'],
+  extends: ["eslint-config-gev/js"],
   ignorePatterns: [],
   rules: {},
-};
+}