+++ examples/react-native-ts.eslintrc.js
@@ -1,25 +1,25 @@
 // https://github.com/SrBrahma/eslint-config-gev
 // This is a workaround for https://github.com/eslint/eslint/issues/3458
-require('@rushstack/eslint-patch/modern-module-resolution');
+require("@rushstack/eslint-patch/modern-module-resolution")
 
 module.exports = {
   root: true,
   env: {
     es2021: true,
     node: true,
-    'react-native/react-native': true, // *1
+    "react-native/react-native": true, // *1
   },
-  extends: ['eslint-config-gev/react-native-js'],
+  extends: ["eslint-config-gev/react-native-js"],
   overrides: [
     {
-      files: ['*.ts', '*.tsx'],
-      extends: ['eslint-config-gev/react-native'],
-      parser: '@typescript-eslint/parser',
+      files: ["*.ts", "*.tsx"],
+      extends: ["eslint-config-gev/react-native"],
+      parser: "@typescript-eslint/parser",
       parserOptions: {
         tsconfigRootDir: __dirname,
-        project: ['./tsconfig.json'],
+        project: ["./tsconfig.json"],
         ecmaVersion: 12,
-        sourceType: 'module',
+        sourceType: "module",
         ecmaFeatures: {
           // *1
           jsx: true,
@@ -28,6 +28,6 @@ module.exports = {
     },
   ],
   rules: {},
-};
+}
 
 // [*1]: https://github.com/Intellicode/eslint-plugin-react-native#configuration