+++ examples/react-ts.eslintrc.js
@@ -1,6 +1,6 @@
 // https://github.com/SrBrahma/eslint-config-gev
 // This is a workaround for https://github.com/eslint/eslint/issues/3458
-require('@rushstack/eslint-patch/modern-module-resolution');
+require("@rushstack/eslint-patch/modern-module-resolution")
 
 module.exports = {
   root: true,
@@ -8,23 +8,23 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ['eslint-config-gev/react-js'],
+  extends: ["eslint-config-gev/react-js"],
   overrides: [
     {
-      files: ['*.ts', '*.tsx'],
-      extends: ['eslint-config-gev/react'],
-      parser: '@typescript-eslint/parser',
+      files: ["*.ts", "*.tsx"],
+      extends: ["eslint-config-gev/react"],
+      parser: "@typescript-eslint/parser",
       parserOptions: {
         tsconfigRootDir: __dirname,
-        project: ['./tsconfig.json'],
+        project: ["./tsconfig.json"],
         ecmaVersion: 12,
-        sourceType: 'module',
+        sourceType: "module",
         ecmaFeatures: {
           jsx: true,
         },
       },
     },
   ],
-  ignorePatterns: ['/lib/**/*', '/dist/**/*'],
+  ignorePatterns: ["/lib/**/*", "/dist/**/*"],
   rules: {},
-};
+}