+++ index.js
@@ -1,5 +1,4 @@
 // Reflects ./ts.js.
-
 module.exports = {
-  extends: ['./ts.js'],
-};
+  extends: ["./ts.js"],
+}