+++ scripts/.eslintrcNoBiome.cjs
@@ -1,5 +1,6 @@
 // This is a workaround for https://github.com/eslint/eslint/issues/3458
 require("@rushstack/eslint-patch/modern-module-resolution")
+const path = require("path")
 
 module.exports = {
   root: true,
@@ -7,22 +8,18 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ["./js.js"],
+  extends: ["../js.js"],
   overrides: [
     {
       files: ["*.ts"],
-      extends: ["./ts.js"],
+      extends: ["../ts.js"],
       parser: "@typescript-eslint/parser",
       parserOptions: {
-        tsconfigRootDir: __dirname,
-        project: ["./tsconfig.json"],
+        tsconfigRootDir: path.resolve(__dirname, ".."),
+        project: ["../tsconfig.json"],
         ecmaVersion: 12,
         sourceType: "module",
       },
-    },
-    {
-      files: ["*.ts", "*.js", "*.tsx", "*.jsx"],
-      extends: ["biome"],
     },
   ],
 }