+++ ts.js
@@ -1,38 +1,39 @@
 module.exports = {
-  plugins: ['@typescript-eslint'],
+  plugins: ["@typescript-eslint"],
   extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/eslint-recommended',
-    'plugin:@typescript-eslint/recommended', // After the others https://stackoverflow.com/a/61555310
-    'plugin:@typescript-eslint/recommended-requiring-type-checking',
-    './js.js',
-    './ts-extensions.js', // Automatic JS to TS rules! I love doing smart stuff! :)
-    './removeFormatter/ts.js',
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended", // After the others https://stackoverflow.com/a/61555310
+    "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "./js.js",
+    "./ts-extensions.js", // Automatic JS to TS rules! I love doing smart stuff! :)
+    "./removeFormatter/ts.js",
+    "biome",
   ],
   rules: {
     // Allows explicitating implicit types, like `const a: number = 10`
-    '@typescript-eslint/no-inferrable-types': 'off',
+    "@typescript-eslint/no-inferrable-types": "off",
 
     // Else would complain about // @ts-ignore in .js files. If I didn't need this ts-comment, this could be removed
-    '@typescript-eslint/ban-ts-comment': 'off',
+    "@typescript-eslint/ban-ts-comment": "off",
 
     // Warn if using a Promise without await/then/catch. Good to avoid running stuff in background non intentionally.
     // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-floating-promises.md
-    '@typescript-eslint/no-floating-promises': 'warn',
+    "@typescript-eslint/no-floating-promises": "warn",
 
     // This wasn't allowing different data types in `templates string`. Why the fuck not?
-    '@typescript-eslint/restrict-template-expressions': 'off',
+    "@typescript-eslint/restrict-template-expressions": "off",
 
     // Wasn't simply allowing `const a = x.y.functionA`.
-    '@typescript-eslint/unbound-method': 'off',
+    "@typescript-eslint/unbound-method": "off",
 
     // [1.1.0] Allows to use `const a = require('...')`. This is specially useful for `const pkg = require('../package.json')`,
     // as TS will complain about importing a file outside rootDir (defined in tsconfig)
-    '@typescript-eslint/no-var-requires': 'off',
+    "@typescript-eslint/no-var-requires": "off",
 
     // Allow having a `function a(callback: () => void) {...}` to be called with a(async () => {...})`
-    '@typescript-eslint/no-misused-promises': [
-      'error',
+    "@typescript-eslint/no-misused-promises": [
+      "error",
       {
         checksVoidReturn: false,
       },
@@ -40,10 +41,10 @@ module.exports = {
 
     // Allow 'aString'.match('a').
     // https://github.com/typescript-eslint/typescript-eslint/blob/v4.28.3/packages/eslint-plugin/docs/rules/prefer-regexp-exec.md
-    '@typescript-eslint/prefer-regexp-exec': 'off',
+    "@typescript-eslint/prefer-regexp-exec": "off",
 
     /** Disable it to allow empty catch blocks. */
-    '@typescript-eslint/no-empty-function': 'off',
+    "@typescript-eslint/no-empty-function": "off",
 
     /**
      * This is kinda useless. It's more of a pain, actually:
@@ -51,11 +52,11 @@ module.exports = {
      * it will mark it as an error.
      * https://eslint.org/docs/rules/no-empty-pattern
      */
-    'no-empty-pattern': 'off',
+    "no-empty-pattern": "off",
 
     /** https://typescript-eslint.io/rules/no-unnecessary-condition/ */
-    '@typescript-eslint/no-unnecessary-condition': [
-      'warn',
+    "@typescript-eslint/no-unnecessary-condition": [
+      "warn",
       {
         allowConstantLoopConditions: true, // Allow `while (true)`
       },
@@ -69,17 +70,17 @@ module.exports = {
     // Allows using `a!.b`. Not safe, but sometimes we really know it is valid.
 
     // Let us have our empty stuff!
-    '@typescript-eslint/no-empty-interface': 'off',
+    "@typescript-eslint/no-empty-interface": "off",
 
     /** [2022-02-03] It's bugging often, so I disabled it. */
-    '@typescript-eslint/restrict-plus-operands': 'off',
+    "@typescript-eslint/restrict-plus-operands": "off",
 
     /**
      * Improves tree-shaking and improves TS optimization.
      * https://github.com/typescript-eslint/typescript-eslint/blob/main/packages/eslint-plugin/docs/rules/consistent-type-imports.md
      */
-    '@typescript-eslint/consistent-type-imports': [
-      'warn',
+    "@typescript-eslint/consistent-type-imports": [
+      "warn",
       {
         disallowTypeAnnotations: false, // Allows import('foo').Bar. Useful in global.d.ts.
       },
@@ -89,8 +90,8 @@ module.exports = {
      * Remove unnecessary boolean literal conditionals`.
      * https://typescript-eslint.io/rules/no-unnecessary-boolean-literal-compare
      */
-    '@typescript-eslint/no-unnecessary-boolean-literal-compare': [
-      'warn',
+    "@typescript-eslint/no-unnecessary-boolean-literal-compare": [
+      "warn",
       {
         allowComparingNullableBooleansToTrue: false,
       },
@@ -100,30 +101,30 @@ module.exports = {
      * TS handles this, it raised false-positives sometimes.
      * https://github.com/Chatie/eslint-config/issues/45#issuecomment-1003990077
      */
-    'no-undef': 'off',
+    "no-undef": "off",
 
     /** -=-=-=- Disallow dangerous stuff, converts recommended warn to error -=-=-=- */
-    '@typescript-eslint/no-explicit-any': 'error',
-    '@typescript-eslint/no-unsafe-assignment': 'error',
-    '@typescript-eslint/no-unsafe-return': 'error',
-    '@typescript-eslint/no-unsafe-member-access': 'error',
-    '@typescript-eslint/no-unsafe-argument': 'error',
-    '@typescript-eslint/no-unsafe-call': 'error',
-    '@typescript-eslint/no-non-null-assertion': 'error',
+    "@typescript-eslint/no-explicit-any": "error",
+    "@typescript-eslint/no-unsafe-assignment": "error",
+    "@typescript-eslint/no-unsafe-return": "error",
+    "@typescript-eslint/no-unsafe-member-access": "error",
+    "@typescript-eslint/no-unsafe-argument": "error",
+    "@typescript-eslint/no-unsafe-call": "error",
+    "@typescript-eslint/no-non-null-assertion": "error",
 
     /**
      * Prefer Array<string> instead of string[]. Faster to read and to change.
      *
      * https://typescript-eslint.io/rules/array-type/
      */
-    '@typescript-eslint/array-type': ['warn', { default: 'generic' }],
+    "@typescript-eslint/array-type": ["warn", { default: "generic" }],
 
     /**
      * https://typescript-eslint.io/rules/prefer-string-starts-ends-with/
      */
-    '@typescript-eslint/prefer-string-starts-ends-with': 'warn',
+    "@typescript-eslint/prefer-string-starts-ends-with": "warn",
   },
-};
+}
 
 // Limbo. Disabled, but may be reenabled at some point.
 // "@typescript-eslint/explicit-module-boundary-types": "off", // Removes the need of exported functions to have explicit rtn type